GroupList working in EditGathering
